Refactor rename UserNotes to Notes
This change updates the import in the index.js layout to better reflect
the purpose of the component, improving code clarity and consistency. It
aligns with the naming convention used for other components. No
significant consequences are expected from this straightforward
renaming.

#ISSUE-123
